backend/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
django-cors-headers==3.2.1
python-dotenv==0.12.0
mysqlclient==1.4.6
gunicorn==20.0.4
django-mysql==3.3.0
mysqlclient==1.4.6
